Move ScreenContent component
