+++ ts/components/screens/MarkdownScreenComponent.tsx
@@ -4,10 +4,10 @@ import { InteractionManager, StyleSheet } from "react-native";
 
 import themeVariables from "../../theme/variables";
 import { ComponentProps } from "../../types/react";
-import ScreenContent from "../shared/ScreenContent";
 import ActivityIndicator from "../ui/ActivityIndicator";
 import Markdown from "../ui/Markdown";
 import BaseScreenComponent from "./BaseScreenComponent";
+import ScreenContent from "./ScreenContent";
 
 const styles = StyleSheet.create({
   contentContainer: {
