+++ ts/components/wallet/TransactionsList.tsx
@@ -21,7 +21,7 @@ import I18n from "../../i18n";
 import { Transaction } from "../../types/pagopa";
 import { cleanTransactionDescription } from "../../utils/payment";
 import { centsToAmount, formatNumberAmount } from "../../utils/stringBuilder";
-import ScreenContent from "../shared/ScreenContent";
+import ScreenContent from "../screens/ScreenContent";
 import { WalletStyles } from "../styles/wallet";
 import BoxedRefreshIndicator from "../ui/BoxedRefreshIndicator";
