+++ ts/screens/development/MarkdownScreen.tsx
@@ -2,7 +2,7 @@ import { Button, Container, Text, View } from "native-base";
 import * as React from "react";
 import { TextInput } from "react-native";
 
-import ScreenContent from "../../components/shared/ScreenContent";
+import ScreenContent from "../../components/screens/ScreenContent";
 import Markdown from "../../components/ui/Markdown";
 import I18n from "../../i18n";
