+++ ts/screens/profile/ProfileMainScreen.tsx
@@ -17,9 +17,9 @@ import { NavigationScreenProp, NavigationState } from "react-navigation";
 import { connect } from "react-redux";
 
 import { withLightModalContext } from "../../components/helpers/withLightModalContext";
+import ScreenContent from "../../components/screens/ScreenContent";
 import TopScreenComponent from "../../components/screens/TopScreenComponent";
 import SelectLogoutOption from "../../components/SelectLogoutOption";
-import ScreenContent from "../../components/shared/ScreenContent";
 import IconFont from "../../components/ui/IconFont";
 import { LightModalContextInterface } from "../../components/ui/LightModal";
 import I18n from "../../i18n";
