+++ ts/screens/wallet/WalletsScreen.tsx
@@ -15,7 +15,7 @@ import WalletLayout from "../../components/wallet/WalletLayout";
 import I18n from "../../i18n";
 
 import { withLoadingSpinner } from "../../components/helpers/withLoadingSpinner";
-import ScreenContent from "../../components/shared/ScreenContent";
+import ScreenContent from "../../components/screens/ScreenContent";
 import { AddPaymentMethodButton } from "../../components/wallet/AddPaymentMethodButton";
 import CardComponent from "../../components/wallet/card/CardComponent";
 import {
